+++ language/diem-framework/releases/Cargo.toml
@@ -16,6 +16,7 @@ diem-types = { path = "../../../types" }
 diem-workspace-hack = { path = "../../../common/workspace-hack" }
 vm = { path = "../../vm" }
 
+bytecode-verifier = {path = "../../bytecode-verifier" }
 anyhow = "1.0.38"
 include_dir = "0.6.0"
 bcs = "0.1.2"
