+++ language/diem-vm/Cargo.toml
@@ -31,6 +31,8 @@ vm = { path = "../vm" }
 serde_json = "1.0.64"
 serde = { version = "1.0.124", default-features = false }
 
+diem-framework-releases = { path = "../diem-framework/releases"}
+
 [dev-dependencies]
 proptest = "1.0.0"
